Rewrite notebooks and mosules
